rename Dealer to MultiSecretRoyAdikari
